requirements.txt
django==3.2.6
djangorestframework==3.12.4
lab-orchestrator-lib==0.0.18
lab-orchestrator-lib-django-adapter==0.0.9
gunicorn==20.0.4
psycopg2==2.9.1
uritemplate==3.0.1
dj-rest-auth==2.1.11
django-allauth==0.45.0
rules==2.2
ai-django-core==4.0.1
django-filter==2.4.0